+++ admin/src/components/common/PropertyShow.tsx
@@ -1,6 +1,8 @@
 import { css } from '@emotion/css';
-import { Stack, Text } from '@keystone-ui/core';
+import { Stack } from '@keystone-ui/core';
 import React, { FC } from 'react';
+import type { ReactNode } from 'react';
+import { Text } from '@chakra-ui/react';
 type DeliveryDetails = {
   direction: string;
   phone: string;
@@ -19,23 +21,19 @@ export const PropertyLine: FC<{
         margin-top: 1rem;
       `}
     >
-      <Text
-        color="black"
-        className={css`
-          font-weight: 600;
-        `}
-      >
-        {label}
-      </Text>
+      <Text mr="2" fontWeight={"semibold"}>{label}</Text>
       <Text marginLeft={'medium'}>{value}</Text>
     </Stack>
   );
 };
 
-export const PropertySection: FC<{ title: string }> = ({ title, children }) => {
+export const PropertySection: FC<{ title: string; children?: ReactNode }> = ({
+  title,
+  children,
+}) => {
   return (
     <Stack padding={'medium'}>
-      <Text size="medium" color="black" weight="bold">
+      <Text size="medium" color="black" fontWeight={"bold"}>
         {title}
       </Text>
       {children}