+++ admin/src/components/order/LineDetails/types.ts
@@ -1,0 +1,11 @@
+import { OrderLineItem } from '.keystone/types';
+
+export type LineItem = OrderLineItem & {
+    id: string;
+    product: {
+      id: string;
+      name: string;
+      price: string;
+    };
+  };
+