+++ admin/src/modules/order/constants.ts
@@ -1,0 +1,4 @@
+export const PATCH_ORDER_COMMANDS = {
+     update_quantity : "UPDATE_QUANTITY",
+     update_line  :  "UPDATE_LINE"
+}