+++ apps/website/src/core/modules/cart/controllers/useUpdateOrder.ts
@@ -5,8 +5,6 @@ import { CreateOrderArgs } from '@smartfood/client/v2';
 export const useUpdateOrder = () => {
   return useMutation(async (args: CreateOrderArgs) => {
     const result = await cms.order.update(args);
-    console.log(result);
-
     return result;
   });
 };