+++ packages/common/src/constants/order.ts
@@ -1,0 +1,4 @@
+export const PATCH_ORDERLINE_COMMANDS = {
+    update_quantity : "UPDATE_QUANTITY",
+    update_line  :  "UPDATE_LINE"
+}