+++ packages/common/src/index.ts
@@ -6,3 +6,4 @@ export * from './whatssApp';
 export * from './matcher';
 export * from './shared-types';
 export * from './shared-constants';
+export * from "./constants/index"