+++ apps/website/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @smartfood/website
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @smartfood/common@1.0.2
+  - @smartfood/client@0.2.2
+  - @smartfood/ui@0.0.4
+
 ## 0.1.3
 
 ### Patch Changes