+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @smartfood/client
 
+## 0.2.2
+
+### Patch Changes
+
+- - Bugs about the update have been solved
+  - The constant `PATCH_ORDERLINE_COMMANDS` has been added
+  - The property command has beed added to `patchOrderLine` to improve performance
+- Updated dependencies
+  - @smartfood/common@1.0.2
+
 ## 0.2.1
 
 ### Patch Changes